Add support for ProxyHost binding in API
Introduced a new `BindExisting` field in `ProxyHostSpec` to allow automatic binding to existing proxy hosts matched by domain names. Updated the CRD to include this field with a default value of true. Implemented deletion of ProxyHost records in the Nginx Proxy Manager only if not bound.

Also, added utility functions to interact with the Nginx Proxy Manager API to find and delete proxy hosts, and initialized related tests for the new functionality.